Separate imports from Icon module
