+++ src/pages/SetPasswordPage.js
@@ -13,7 +13,8 @@ import lodashHas from 'lodash.has';
 import compose from '../libs/compose';
 import {Redirect, withRouter} from '../libs/Router';
 import styles from '../styles/styles';
-import Icon, {BRAND_ASSETS} from '../components/Icon';
+import Icon from '../components/Icon';
+import BRAND_ASSETS from '../components/Icon/BRAND_ASSETS';
 import CustomStatusBar from '../components/CustomStatusBar';
 import {setPassword} from '../libs/actions/Session';
 import ONYXKEYS from '../ONYXKEYS';
