+++ src/pages/home/sidebar/ChatSwitcherSearchForm.js
@@ -10,7 +10,8 @@ import PropTypes from 'prop-types';
 import styles from '../../../styles/styles';
 import themeColors from '../../../styles/themes/default';
 import variables from '../../../styles/variables';
-import Icon, {BRAND_ASSETS} from '../../../components/Icon';
+import Icon from '../../../components/Icon';
+import BRAND_ASSETS from '../../../components/Icon/BRAND_ASSETS';
 import TextInputWithFocusStyles from '../../../components/TextInputWithFocusStyles';
 import iconX from '../../../../assets/images/icon-x.png';
 import {getDisplayName} from '../../../libs/actions/PersonalDetails';
